Test if file exist, before deleting it, to eliminate error msg

+++ www/lib/sadm_footer.php
@@ -17,7 +17,7 @@
     pg_close($connection);
 	
 	# Delete Temporary File define in sadm_init.php
-	unlink (SADM_WWW_TMP_FILE1);
+	if (file_exists($filename)) { unlink (SADM_WWW_TMP_FILE1); }
 	
  ?>
